// src/app/models/Book.model.ts
export default class AdzBook {
    constructor() { }
    title: string;
    authors: string[];
    imageLink: string;
    country: string;
    description: string
    language: string;
    link: string;
    pageCount: number;
    publishedDate: any;
    publisher: string;
    categories: string[];
    price: number;
    currency: string;
    isEbook: Boolean;
}